+++ Constraint/ConstraintParser.h
@@ -24,6 +24,7 @@ int create_new_constraint(const shared_ptr<DomainBase>&, istringstream&);
 
 void new_embed2d(unique_ptr<Constraint>&, istringstream&);
 void new_fixedlength(unique_ptr<Constraint>&, istringstream&, unsigned);
+void new_minimumgap(unique_ptr<Constraint>&, istringstream&, unsigned);
 void new_mpc(unique_ptr<Constraint>&, istringstream&);
 void new_particlecollision2d(unique_ptr<Constraint>&, istringstream&);
 void new_particlecollision3d(unique_ptr<Constraint>&, istringstream&);